Update and rename Balance.py to Reaction_Balance.py
Additional functionality added. Balance will be done with whole number coefficients by default. Some lists were replaced with dictionaries for convenience. Basis compound feature added, which allows for the user to preset one compound to be balanced with a coefficient of 1 (for example, a fuel). Comments have been added to