Updated Reaction Handling
The reactions section of the control volumes is now handled similarly to the info equations. A general dictionary with every reaction present is saved to be used in different control volumes. When defining a control volume, the user will be prompted to state which reactions (using the integer that represents the order they were added, starting at 1) are present. Any chemicals in the reaction will automatically be added to the streams of the volume. More testing probably needs to be done to make sure it works properly (I foresee issues with extent of reaction not substituting properly).