Updated Reaction Handling
The setup() function here now requires an argument of all the reactions in a dictionary (the output of the reaction setup() function). This format should save the user a lot of inputs, a problem that I won't really get to fixing until I figure out why Qt Designer doesn't work on my computers.