Updated files to use OOP
Using classes and PyQt I've been able to create a usable user interface and much more efficient code. There is still a lot I can do with this project, as shown in the To-Do list in my UI file comments, but for now I may run a few more tests and call it good. I need to get started on Energy Balances and give this code a bit of a break. It's come a long way.