New Reaction Object
This is used to more conveniently store information about a chemical reaction. It can handle compounds and their coefficients, extent of reaction and heat of reaction so far. It will check itself to see if it is balanced and balance the reaction if needed (defaulted to whole coefficients). If one desires a compound be kept at 1 mole (for example, a fuel) then a method has been included to accommodate for this as well.